Move the database models into the db submodule
This continues the effort to clean up the database-related code to
avoid circular imports by moving the database models, session and
initialization code into the db subpackage. This should have no
functional change.
