+++ anitya/tests/db/test_models.py
@@ -30,125 +30,11 @@ import unittest
 from sqlalchemy.dialects import postgresql, sqlite
 from sqlalchemy.types import CHAR
 from sqlalchemy.exc import IntegrityError
-import mock
 import six
 
 import anitya.lib.model as model
 from anitya.lib import versions
 from anitya.tests.base import DatabaseTestCase, create_distro, create_project, create_package
-
-
-class InitalizeTests(unittest.TestCase):
-
-    @mock.patch('anitya.lib.model.sa.create_engine')
-    @mock.patch('anitya.lib.model.Session')
-    def test_initialize(self, mock_session, mock_create_engine):
-        config = {'DB_URL': 'postgresql://postgres:pass@localhost/mydb'}
-        engine = model.initialize(config)
-        mock_create_engine.assert_called_once_with(config['DB_URL'], echo=False)
-        self.assertEqual(engine, mock_create_engine.return_value)
-        mock_session.configure.assert_called_once_with(bind=engine)
-
-    @mock.patch('anitya.lib.model.sa.create_engine')
-    @mock.patch('anitya.lib.model.sa.event.listen')
-    @mock.patch('anitya.lib.model.Session')
-    def test_initalize_sqlite(self, mock_session, mock_listen, mock_create_engine):
-        config = {'DB_URL': 'sqlite://', 'SQL_DEBUG': True}
-        engine = model.initialize(config)
-        mock_create_engine.assert_called_once_with(config['DB_URL'], echo=True)
-        mock_session.configure.assert_called_once_with(bind=engine)
-        self.assertEqual(1, mock_listen.call_count)
-        self.assertEqual(engine, mock_listen.call_args_list[0][0][0])
-        self.assertEqual('connect', mock_listen.call_args_list[0][0][1])
-
-
-class BaseQueryPaginateTests(DatabaseTestCase):
-    """Tests for the BaseQuery queries."""
-
-    def setUp(self):
-        super(BaseQueryPaginateTests, self).setUp()
-        self.query = model.BaseQuery(model.Project, session=self.session)
-
-    def test_defaults(self):
-        """Assert paginate defaults to the first page and 25 items."""
-        create_project(self.session)
-        page = self.query.paginate()
-        self.assertEqual(1, page.page)
-        self.assertEqual(3, page.total_items)
-        self.assertEqual(25, page.items_per_page)
-        # Default ordering is just by id
-        self.assertEqual(page.items[0].name, 'geany')
-        self.assertEqual(page.items[1].name, 'subsurface')
-        self.assertEqual(page.items[2].name, 'R2spec')
-
-    def test_multiple_pages(self):
-        """Assert multiple pages work with pagination."""
-        create_project(self.session)
-        page = self.query.paginate(items_per_page=2)
-        page2 = self.query.paginate(page=2, items_per_page=2)
-        self.assertEqual(1, page.page)
-        self.assertEqual(3, page.total_items)
-        self.assertEqual(2, page.items_per_page)
-        self.assertEqual(page.items[0].name, 'geany')
-        self.assertEqual(page.items[1].name, 'subsurface')
-
-        self.assertEqual(2, page2.page)
-        self.assertEqual(3, page2.total_items)
-        self.assertEqual(2, page2.items_per_page)
-        self.assertEqual(page2.items[0].name, 'R2spec')
-
-    def test_no_results(self):
-        """Assert an empty page is returned when page * items_per_page > total_items."""
-        create_project(self.session)
-        page = self.query.paginate(page=1000)
-        self.assertEqual(1000, page.page)
-        self.assertEqual(3, page.total_items)
-        self.assertEqual(25, page.items_per_page)
-        self.assertEqual(0, len(page.items))
-
-    def test_nonsense_page(self):
-        """Assert a page number less than 1 raises a ValueError."""
-        self.assertRaises(ValueError, self.query.paginate, 0)
-        self.assertRaises(ValueError, self.query.paginate, -1)
-
-    def test_nonsense_items_per_page(self):
-        """Assert an items_per_page number less than 1 raises a ValueError."""
-        self.assertRaises(ValueError, self.query.paginate, 1, 0)
-        self.assertRaises(ValueError, self.query.paginate, 1, -1)
-
-    def test_order_by(self):
-        """Assert you can alter the order of page results."""
-        create_project(self.session)
-        page = self.query.paginate(order_by=model.Project.name)
-        self.assertEqual(1, page.page)
-        self.assertEqual(3, page.total_items)
-        self.assertEqual(25, page.items_per_page)
-        self.assertEqual(page.items[0].name, 'R2spec')
-        self.assertEqual(page.items[1].name, 'geany')
-        self.assertEqual(page.items[2].name, 'subsurface')
-
-    def test_as_dict(self):
-        expected_dict = {
-            u'items_per_page': 1,
-            u'page': 1,
-            u'total_items': 3,
-            u'items': [{
-                'id': 3,
-                'backend': u'custom',
-                'name': u'R2spec',
-                'homepage': u'https://fedorahosted.org/r2spec/',
-                'regex': None,
-                'version': None,
-                'version_url': None,
-                'versions': [],
-            }],
-        }
-        create_project(self.session)
-        page = self.query.paginate(order_by=model.Project.name, items_per_page=1)
-        actual_dict = page.as_dict()
-        actual_dict['items'][0].pop('updated_on')
-        actual_dict['items'][0].pop('created_on')
-        self.assertEqual(expected_dict, actual_dict)
 
 
 class ProjectTests(DatabaseTestCase):
